+++ bp_search/templates/search/search_home.html
